refactored UsersServiceImpl and removed all follower logic into its own FollowerServiceImpl class
