+++ CarnasaGamesWebsiteAndApi/src/main/java/com/sparta/cr/carnasagameswebsiteandapi/controllers/api/FollowerApiController.java
@@ -1,7 +1,7 @@
 package com.sparta.cr.carnasagameswebsiteandapi.controllers.api;
 
 import com.sparta.cr.carnasagameswebsiteandapi.models.FollowerModel;
-import com.sparta.cr.carnasagameswebsiteandapi.models.UserModel;
+import com.sparta.cr.carnasagameswebsiteandapi.services.implementations.FollowerServiceImpl;
 import com.sparta.cr.carnasagameswebsiteandapi.services.implementations.UserServiceImpl;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.hateoas.CollectionModel;
@@ -10,37 +10,37 @@ import org.springframework.hateoas.Link;
 import org.springframework.hateoas.server.mvc.WebMvcLinkBuilder;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
-import java.util.stream.Collectors;
 
 @RestController
 @RequestMapping("/api/users")
 public class FollowerApiController {
 
+    private final FollowerServiceImpl followerService;
     private final UserServiceImpl userService;
 
     @Autowired
-    public FollowerApiController(UserServiceImpl userService) {
+    public FollowerApiController(FollowerServiceImpl followerService, UserServiceImpl userService) {
+        this.followerService = followerService;
         this.userService = userService;
     }
 
     @GetMapping("/search/id/{userId}/following")
     public ResponseEntity<CollectionModel<EntityModel<FollowerModel>>> getAllFollowingByUserId(@PathVariable("userId") Long userId){
-        List<EntityModel<FollowerModel>> following = userService.getAllFollowingByUserId(userId).stream().map( followerModel ->
+        List<EntityModel<FollowerModel>> following = followerService.getAllFollowingByUserId(userId).stream().map( followerModel ->
                 getUserEntityModel(followerModel).add(getUserLink(followerModel))
         ).toList();
-        return ResponseEntity.ok(CollectionModel.of(following));
+        return ResponseEntity.ok(CollectionModel.of(following).add(WebMvcLinkBuilder.linkTo(WebMvcLinkBuilder.methodOn(UserApiController.class).getUserById(userId)).withSelfRel()));
     }
 
     @GetMapping("/search/id/{userId}/followers")
     public ResponseEntity<CollectionModel<EntityModel<FollowerModel>>> getAllFollowersByUserId(@PathVariable("userId") Long userId){
-        List<EntityModel<FollowerModel>> followers = userService.getAllFollowersByUserId(userId).stream().map( followerModel ->
+        List<EntityModel<FollowerModel>> followers = followerService.getAllFollowersByUserId(userId).stream().map( followerModel ->
                 getUserEntityModel(followerModel).add(getFollowerLink(followerModel))
         ).toList();
-        return ResponseEntity.ok(CollectionModel.of(followers));
+        return ResponseEntity.ok(CollectionModel.of(followers).add(WebMvcLinkBuilder.linkTo(WebMvcLinkBuilder.methodOn(UserApiController.class).getUserById(userId)).withSelfRel()));
     }
 
     @GetMapping("/search/id/{userId}/followers/number")
@@ -48,13 +48,13 @@ public class FollowerApiController {
         if(userService.getUser(userId).isEmpty()){
             return new ResponseEntity<>(HttpStatus.NOT_FOUND);
         }
-        return ResponseEntity.ok(userService.getNumberOfFollowersByUserId(userId));
+        return ResponseEntity.ok(followerService.getNumberOfFollowersByUserId(userId));
     }
 
     @PostMapping("/new/follower")
     public ResponseEntity<EntityModel<FollowerModel>> followNewUser(@RequestBody FollowerModel followerModel){
-        userService.validateNewFollower(followerModel);
-        userService.followNewUser(followerModel);
+        followerService.validateNewFollower(followerModel);
+        followerService.followNewUser(followerModel);
         return new ResponseEntity<>(HttpStatus.CREATED);
     }
 
@@ -63,7 +63,7 @@ public class FollowerApiController {
         if(userService.getUser(userId).isEmpty()||userService.getUser(followerId).isEmpty()){
             return new ResponseEntity<>(HttpStatus.NOT_FOUND);
         }
-        userService.unfollowUser(userId, followerId);
+        followerService.unfollowUser(userId, followerId);
         return new ResponseEntity<>(HttpStatus.NO_CONTENT);
     }
